use print_errno_error in redirection
