+++ includes/errors.h
@@ -43,7 +43,7 @@ t_error const 		*get_error_defs(void);
 void				print_error_msg(t_error_id id);
 void				print_usage_msg(t_builtin_id id);
 void				print_name_and_error(t_error_id id);
-void 				print_file_error(t_error_id id, char *file);
+void				print_errno_error(int err, char const *pre, char const *post);
 
 t_error_id			get_error(void);
 void				set_error(t_error_id error);
